soundsystem/src/test/java/TrackCounterTest.java
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
import root.aspects.TrackCounter;
import root.config.TrackConuntingConfig;
import root.soundsystem.CompactDisc;

import static org.junit.Assert.assertEquals;

@RunWith(SpringJUnit4ClassRunner.class)
@ContextConfiguration(classes = TrackConuntingConfig.class)
public class TrackCounterTest {

    @Autowired
    private CompactDisc cd;

    @Autowired
    private TrackCounter trackCounter;

    @Test
    public void testTrackCounter(){
        cd.playTrack(1);
        cd.playTrack(2);
        cd.playTrack(2);
        cd.playTrack(2);

        assertEquals((Integer) 1,trackCounter.getPlayCount(1));
        assertEquals((Integer) 3,trackCounter.getPlayCount(2));
    }

}
